requirements.txt
ultralytics>=8.3.0
streamlit
torch
torchvision
onnx
onnxruntime
opencv-python-headless
numpy
pillow
